requirements.txt
alembic==1.12.1
apache-airflow==1.10.10
apispec==1.3.3
argcomplete==1.12.3
attrs==19.3.0
Babel==2.13.1
blinker==1.6.3
cached-property==1.5.2
cachetools==4.2.4
cattrs==0.9.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.2
click==7.1.2
colorama==0.4.6
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.37
db-dtypes==1.1.1
defusedxml==0.7.1
dill==0.3.7
dnspython==2.3.0
docutils==0.20.1
email-validator==1.3.1
Flask==1.1.4
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.13
Flask-WTF==0.14.3
funcsigs==1.0.2
future==0.18.3
google-api-core==1.34.0
google-api-python-client==1.12.11
google-auth==1.35.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==0.5.3
google-cloud-bigquery==2.34.4
google-cloud-bigquery-storage==2.22.0
google-cloud-bigtable==2.21.0
google-cloud-container==2.33.0
google-cloud-core==2.3.3
google-cloud-dlp==3.13.0
google-cloud-language==2.11.1
google-cloud-secret-manager==2.16.4
google-cloud-spanner==3.40.1
google-cloud-speech==2.22.0
google-cloud-storage==2.11.0
google-cloud-texttospeech==2.14.2
google-cloud-translate==3.12.1
google-cloud-videointelligence==2.11.4
google-cloud-vision==3.4.5
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
graphviz==0.20.1
greenlet==3.0.1
grpc-google-iam-v1==0.12.7
grpcio==1.59.2
grpcio-gcp==0.2.2
grpcio-status==1.48.2
gunicorn==19.10.0
httplib2==0.22.0
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.12.0
iso8601==2.1.0
itsdangerous==1.1.0
Jinja2==2.10.3
json-merge-patch==0.2
jsonschema==3.2.0
lazy-object-proxy==1.9.0
lockfile==0.12.2
Mako==1.2.4
Markdown==2.6.11
MarkupSafe==2.0.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
natsort==8.4.0
numpy==1.20.0
oauthlib==3.2.2
packaging==21.3
pandas==0.25.3
pandas-gbq==0.14.1
pendulum==1.4.4
prison==0.2.1
proto-plus==1.22.3
protobuf==3.20.3
psutil==5.9.6
py4j==0.10.7
pyarrow==7.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydata-google-auth==1.8.2
Pygments==2.16.1
PyJWT==1.7.1
pyOpenSSL==23.3.0
pyparsing==3.1.1
pyrsistent==0.19.3
pyspark==2.4.5
python-daemon==2.1.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sentry-sdk==1.34.0
setproctitle==1.3.3
six==1.16.0
SQLAlchemy==1.3.23
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
statsd==3.3.0
tabulate==0.8.10
tenacity==4.12.0
termcolor==1.1.0
text-unidecode==1.2
thrift==0.16.0
typing_extensions==4.7.1
tzlocal==1.5.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==2.0.7
Werkzeug==0.16.1
WTForms==2.3.3
zipp==3.15.0
zope.deprecation==4.4.0
